fix: correctly handle multi-root formatter settings
Previous implementation did create a static Settings.XamlStyler on every update of the VSCode (xaml styler) configuration. This resultet in always the latest updated configuration to be used for all different multi-root projects. The new implementation collects the effective VSCode settings for the specific file and then creates the Settings.XamlStyler config file (temporarely) on the fly.